+++ openbb_terminal/portfolio/portfolio_analysis/reportlab_helpers.py
