+++ openbb_terminal/portfolio/portfolio_controller.py
@@ -207,16 +207,16 @@ class PortfolioController(BaseController):
             "--raw": {},
         }
         choices["distr"] = {
-            "--period": {c: {} for c in portfolio_helper.PERIODS},
+            "--period": {c: {} for c in statics.PERIODS},
             "-p": "--period",
             "--raw": {},
         }
         choices["rvol"] = {
-            "--period": {c: {} for c in portfolio_helper.PERIODS},
+            "--period": {c: {} for c in statics.PERIODS},
             "-p": "--period",
         }
         r_auto_complete = {
-            "--period": {c: {} for c in portfolio_helper.PERIODS},
+            "--period": {c: {} for c in statics.PERIODS},
             "-p": "--period",
             "--rfr": None,
             "-r": "--rfr",
@@ -224,7 +224,7 @@ class PortfolioController(BaseController):
         choices["rsharpe"] = r_auto_complete
         choices["rsort"] = r_auto_complete
         choices["rbeta"] = {
-            "--period": {c: {} for c in portfolio_helper.PERIODS},
+            "--period": {c: {} for c in statics.PERIODS},
             "-p": "--period",
         }
         choices["alloc"] = {c: {} for c in self.AGGREGATION_METRICS}
@@ -454,12 +454,12 @@ class PortfolioController(BaseController):
                 f"\n[bold][param]Portfolio:[/param][/bold] {self.portfolio_name}"
             )
 
-            self.benchmark_name = "SPDR S&P 500 ETF Trust (SPY)"
+            self.risk_free_rate = ns_parser.risk_free_rate / 100
             console.print(
                 f"[bold][param]Risk Free Rate:[/param][/bold] {self.risk_free_rate:.2%}"
             )
 
-            self.risk_free_rate = ns_parser.risk_free_rate / 100
+            self.benchmark_name = "SPDR S&P 500 ETF Trust (SPY)"
             console.print(
                 f"[bold][param]Benchmark:[/param][/bold] {self.benchmark_name}\n"
             )
@@ -622,7 +622,7 @@ class PortfolioController(BaseController):
             "-p",
             "--period",
             type=str,
-            choices=portfolio_helper.PERIODS,
+            choices=statics.PERIODS,
             dest="period",
             default="all",
             help="Period in which to calculate attribution",
@@ -1390,7 +1390,7 @@ class PortfolioController(BaseController):
             "-p",
             "--period",
             type=str,
-            choices=portfolio_helper.PERIODS,
+            choices=statics.PERIODS,
             dest="period",
             default="all",
             help="The file to be loaded",
@@ -1428,7 +1428,7 @@ class PortfolioController(BaseController):
             "-p",
             "--period",
             type=str,
-            choices=portfolio_helper.PERIODS,
+            choices=statics.PERIODS,
             dest="period",
             default="all",
             help="The file to be loaded",
